+++ src/war3/text.lua
@@ -3,7 +3,7 @@ local ej = require 'war3.enhanced_jass'
 local Math = require 'std.math'
 
 local Text = require 'std.class'('Text')
-local Init, SetMode, Scaling, MoveUp, MoveDown, MoveLeft, MoveRight, MoveSin, MoveCos
+local Init, SetTrace, Scaling, Move, MoveSin, ComputeSinTrace, ComputeMin, UpdateText
 
 -- ex: Text:new{
 --     text: 文字,
@@ -11,14 +11,14 @@ local Init, SetMode, Scaling, MoveUp, MoveDown, MoveLeft, MoveRight, MoveSin, Mo
 --     time: 秒數,
 --     mode: 移動模式,
 --     font_size: {字體最小值, 字體增大值, 字體最大值},
---     height: {基本高度, 增加高度(沒用填0)},
+--     height: {基本高度, 增加高度(沒用填0), 最大高度},
 --     offset(可選): Point(x偏移量, y偏移量),
 -- }
 function Text:_new(data)
     data._invalid_ = false
     data._object_ = ej.CreateTextTag()
     data.init = Init
-    SetMode(data)
+    SetTrace(data)
 
     return data
 end
@@ -27,7 +27,7 @@ Init = function(self)
     local TIME_FADE = 0.3
 
     ej.SetTextTagText(self._object_, self.text, self.font_size[1])
-    ej.SetTextTagPos(self._object_, self.loc.x, self.loc.y, self.font_size[1] * self.height[1])
+    ej.SetTextTagPos(self._object_, self.loc.x, self.loc.y, self.height[1])
 
     -- 設置結束點、淡出動畫時間
     ej.SetTextTagPermanent(self._object_, self.time > 0 and false or true)
@@ -35,43 +35,55 @@ Init = function(self)
     ej.SetTextTagFadepoint(self._object_, TIME_FADE)
 end
 
-SetMode = function(self)
-    if self.mode == "fix" then  -- 原地縮放
+SetTrace = function(self)
+    if self.mode == 'fix' then -- 原地縮放
         self.update = Scaling
-    elseif self.mode == "up" then  -- 向上移動
-        self.update = MoveUp
-    elseif self.mode == "down" then  -- 向下移動
-        self.update = MoveDown
-    elseif self.mode == "left" then  -- 向左移動
-        self.update = MoveLeft
-    elseif self.mode == "right" then  -- 向右移動
-        self.update = MoveRight
-    elseif self.mode == "sin" then  -- 正弦移動
+    elseif self.mode == 'move' then -- 移動
+        self.update = Move
+    elseif self.mode == 'sin' then -- 正弦移動
         self.update = MoveSin
-    elseif self.mode == "cos" then  -- 餘弦移動
-        self.update = MoveCos
     end
 end
 
-Scaling = function(self)
+Scaling = function(self, runtime)
+    ej.SetTextTagText(self._object_, self.text, ComputeMin(self.font_size, runtime))
 end
 
-MoveUp = function(self)
+Move = function(self, runtime)
+    -- 更新位置
+    local displacement = runtime * self.offset
+    local loc = self.loc + displacement
+
+    UpdateText(self, loc, ComputeMin(self.font_size, runtime),
+               ComputeMin(self.height, runtime))
+
+    loc:remove()
+    displacement:remove()
 end
 
-MoveDown = function(self)
+ComputeMin = function(x, t)
+    return Math.min(x[1] + x[2] * t, x[3])
 end
 
-MoveLeft = function(self)
+MoveSin = function(self, runtime)
+    -- 更新位置
+    local displacement = runtime * self.offset
+    local loc = self.loc + displacement
+
+    UpdateText(self, loc, ComputeSinTrace(self.font_size, self.time, runtime),
+               ComputeSinTrace(self.height, self.time, runtime))
+
+    loc:remove()
+    displacement:remove()
 end
 
-MoveRight = function(self)
+ComputeSinTrace = function(x, max, t)
+    return (x[3]-x[1]) * Math.sin(Math.pi * t / max) + x[1]
 end
 
-MoveSin = function(self)
-end
-
-MoveCos = function(self)
+UpdateText = function(self, loc, font_size, height)
+    ej.SetTextTagText(self._object_, self.text, font_size)
+    ej.SetTextTagPos(self._object_, loc.x, loc.y, height)
 end
 
 function Text:_remove()
@@ -84,7 +96,7 @@ function Text:_remove()
 end
 
 function Text:start()
-    local PERIOD = 0.1
+    local PERIOD = 0.04
 
     self:init()
 
@@ -93,9 +105,7 @@ function Text:start()
         PERIOD,
         (self.time > 0) and self.time / PERIOD or -1,
         function(timer)
-            if self.update then
-                self:update()
-            end
+            self:update(timer:getRuntime())
 
             if self._invalid_ then
                 self._timer_:stop()